backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.2
dnspython==2.7.0
filelock==3.18.0
flask==3.1.0
flask-cors==6.0.0
flask-pymongo==3.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
pycparser==2.22
pycryptodome==3.23.0
pyjwt==2.10.1
pymongo==4.12.1
python-dotenv==1.1.0
requests==2.32.3
requests-file==2.1.0
tldextract==5.3.0
urllib3==2.4.0
validators==0.35.0
werkzeug==3.1.3
